Add months to Strategy Lab period selector, fix abort timeout
Period options now include 1M/3M/6M alongside 1Y-20Y
Backend accepts float lookback_years (was Literal int enum)
Date math uses timedelta instead of year subtraction
Timeout scales with period (min 3min, 0.5min per year for long runs)
LookbackYears type widened to number

+++ backend/app/core/backtest_engine.py
@@ -380,7 +380,7 @@ def _calculate_metrics(
     trades: list[TradeResult],
     starting_capital: float,
     final_equity: float,
-    lookback_years: int,
+    lookback_years: float,
 ) -> BacktestMetrics:
     if not trades:
         return BacktestMetrics(
